% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{Compute_S_n}
\alias{Compute_S_n}
\title{Compute_S_n_pois
Compute the test statistic of the LT based GOF when the distance is
the integrated squared difference between the empirical Laplace transform
and the theoretical Laplace transform}
\usage{
Compute_S_n(X_data, X_par, N_dist, U_dist, beta, scaled)
}
\arguments{
\item{X_data}{Sample of aggregated losses}

\item{X_par}{Parameter of the compound distribution}

\item{N_dist}{Claim frequency distribution c('pois', 'geom')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', 'invgauss')}

\item{beta}{Parameter of the weight function}

\item{scaled}{boolean indicating wether a scale is applied to the data or not}
}
\value{
The value of the test statistic
}
\description{
Compute_S_n_pois
Compute the test statistic of the LT based GOF when the distance is
the integrated squared difference between the empirical Laplace transform
and the theoretical Laplace transform
}
