% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{Compute_T_n}
\alias{Compute_T_n}
\title{Compute_T_n
Compute the test statistics for the KF-Gamma model based on an ODE}
\usage{
Compute_T_n(X_data, X_par, N_dist, U_dist, beta, scaled)
}
\arguments{
\item{X_data}{Sample of aggregated losses}

\item{X_par}{Parameter of the compound distribution}

\item{N_dist}{Distribution if the claim sizes c('pois', 'geom')}

\item{U_dist}{Distribution if the claim sizes c('exp', 'gamma', invgauss)}

\item{beta}{Parameter in the weight function}

\item{scaled}{Indicates wether or not the data is being scaled}
}
\value{
The value of the test statistic
}
\description{
Compute_T_n
Compute the test statistics for the KF-Gamma model based on an ODE
}
