% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_gof_CD.r
\name{cvm_gof}
\alias{cvm_gof}
\title{CvM_gof
Goodness-of-fit-test based on the Cramer-von Mises criterium}
\usage{
cvm_gof(X_data, N_dist, U_dist, par_method_proxy, method_proxy_df, B,
  method = "partial-MME")
}
\arguments{
\item{X_data}{Samle of aggregated losses}

\item{N_dist}{Claim frequency distribution under H0}

\item{U_dist}{Claim sizes distribution under H0}

\item{par_method_proxy}{Width of the discretization of the claim sizes distribution
when Panjer algorithm is used, order of truncation when the truncation method is used}

\item{method_proxy_df}{Appproximation method for the df of the compound distribution}

\item{B}{Number of bootstrap loops}

\item{method}{MM estimation method c('MME', 'partial-MME')}
}
\value{
Value of the test statistic and critical value of the test
}
\description{
CvM_gof
Goodness-of-fit-test based on the Cramer-von Mises criterium
}
