% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_CD.r
\name{df_X}
\alias{df_X}
\title{df_X
Distribution function of the aggregated losses approximated via Panjer algorithm}
\usage{
df_X(X_par, N_dist, U_dist, max_X = NULL, par_method_proxy, method_proxy)
}
\arguments{
\item{X_par}{Parameter of the compound distribution}

\item{N_dist}{Claim frequency distribution c('pois', 'geom')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', 'weibull', 'invgauss')}

\item{max_X}{Truncation above the threshold max_X}

\item{par_method_proxy}{Width of the discretization of the claim sizes distribution
when Panjer algorithm is used, order of truncation when the truncation method is used}

\item{method_proxy}{Aproximation method for the compound distribution DF c('recursive', 'truncation')}
}
\value{
The distribution function of X as a function
}
\description{
df_X
Distribution function of the aggregated losses approximated via Panjer algorithm
}
