% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{dLT_gamma}
\alias{dLT_gamma}
\title{dLT_gamma
First derivative of the Laplace transform of the gamma distribution}
\usage{
dLT_gamma(t, r, theta)
}
\arguments{
\item{t}{Point at which the derivative of the Laplace transform is evaluated}

\item{r}{Shape parameter of the gamma distribution}

\item{theta}{Scale paameter of the gamma distribution}
}
\value{
the value of the first derivative of the Laplace transform of a gamma distribuytion at t
}
\description{
dLT_gamma
First derivative of the Laplace transform of the gamma distribution
}
