% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{dLT_invgauss}
\alias{dLT_invgauss}
\title{dLT_invgauss
First derivative of the Laplace transform of the inverse Gaussian distribution}
\usage{
dLT_invgauss(t, mu, phi)
}
\arguments{
\item{t}{point at which the first derivative of the Laplace transform of the inverse
Gaussian distribution is evaluated}

\item{mu}{Mean parameter of the inverse Gaussian distribution}

\item{phi}{Dispersion parameter of the inverse Gausian distribution}
}
\value{
The value of first derivative of the Laplace transform of the inverse
Gaussian distribution at t
}
\description{
dLT_invgauss
First derivative of the Laplace transform of the inverse Gaussian distribution
}
