% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTCD-package.R
\docType{package}
\name{FTCD}
\alias{FTCD}
\alias{FTCD-package}
\title{mypkgr.}
\description{
This the place to speak about your package.
}
\details{
What is in the package...

What are the most important functions...
}
