% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{K_1_gamma}
\alias{K_1_gamma}
\title{K_1_gamma
Function for the LT based goodness of fit test for the KF-gamma distribution}
\usage{
K_1_gamma(x, r, theta, a)
}
\arguments{
\item{x}{Point at which the function is evaluated}

\item{r}{Shape paremeter of the gamma distribution}

\item{theta}{Scale parameter of the gamma distribution}

\item{a}{Parameter of the claim frequency distribution}
}
\value{
The value of the function K_1 at x
}
\description{
K_1_gamma
Function for the LT based goodness of fit test for the KF-gamma distribution
}
