% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{K_1_invgauss}
\alias{K_1_invgauss}
\title{K_1_invgauss
Function for the evaluation of test statistic for the LT based test using an ODE
as a distance when testing a KF(a,b)-invgauss(mu, phi) model}
\usage{
K_1_invgauss(x, mu, phi, a)
}
\arguments{
\item{x}{point at which the function K_1 is evaluated}

\item{mu}{Mean parameter of the inverse gaussian distribution governing the claim sizes under
H0}

\item{phi}{Dispersion parameter of the inverse gaussian distribution governing the
claim sizes under H0}

\item{a}{Parameter of the claim frequency distribution}
}
\value{
the value of the function at any given point
}
\description{
K_1_invgauss
Function for the evaluation of test statistic for the LT based test using an ODE
as a distance when testing a KF(a,b)-invgauss(mu, phi) model
}
