% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{K_2_gamma}
\alias{K_2_gamma}
\title{K_2_gamma
Function to evaluate the test statistic in the LT based GOF test with a distance defined as an ODE for the
 KF-Gamma model}
\usage{
K_2_gamma(x, r, theta, a)
}
\arguments{
\item{x}{Point at which the function is evaluated}

\item{r}{Shape parameter of the gamma distribution}

\item{theta}{Scale parameter of the gamma distribution}

\item{a}{Parameter of the claim frequency distribution}
}
\value{
the value of the function at x
}
\description{
K_2_gamma
Function to evaluate the test statistic in the LT based GOF test with a distance defined as an ODE for the
 KF-Gamma model
}
