% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{K_2_invgauss}
\alias{K_2_invgauss}
\title{K_2_invgauss
Function for the evaluation of the test statistic associated to the LT based goodness of fit test for the
KF(a,b)-Inverse Gaussian model}
\usage{
K_2_invgauss(x, mu, phi, a)
}
\arguments{
\item{x}{Point at which the function is evaluated}

\item{mu}{Mean parameter of the inverse gaussian distribution}

\item{phi}{Dispersion parameter of the inverse gaussian distribution}

\item{a}{Parameter of the claim frequency distribution}
}
\value{
The value of the funtion at x
}
\description{
K_2_invgauss
Function for the evaluation of the test statistic associated to the LT based goodness of fit test for the
KF(a,b)-Inverse Gaussian model
}
