% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{LT_gamma}
\alias{LT_gamma}
\title{LT_gamma
Laplace transform of the gamma distribution}
\usage{
LT_gamma(t, r, theta)
}
\arguments{
\item{t}{point at which the Laplace transform is evaluated}

\item{r}{Shape parameter of the gamma distribution}

\item{theta}{Scale parameter of the gamma distribution}
}
\value{
The value of the Laplace transform at t
}
\description{
LT_gamma
Laplace transform of the gamma distribution
}
