% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{LT_invgauss}
\alias{LT_invgauss}
\title{LT_invgauss
Laplace transform of the inverse gaussian distribution}
\usage{
LT_invgauss(t, mu, phi)
}
\arguments{
\item{t}{point at which the Laplace transform is evaluated}

\item{mu}{mean parameter of the inverse gaussian distribution}

\item{phi}{dispersion parameter of the inverse gaussian}
}
\value{
Value of the Laplace transform at t
}
\description{
LT_invgauss
Laplace transform of the inverse gaussian distribution
}
