% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{lt_L2_gof}
\alias{lt_L2_gof}
\title{lt_L2_gof_pois
Compute the test statistic and critical value via a parametric bootstrap routine
for the LT based GOF test with a distance defined as an integrated square differrence}
\usage{
lt_L2_gof(X_data, N_dist, U_dist, beta, B, method = "partial-MME",
  scaled)
}
\arguments{
\item{X_data}{Sample of aggregated losses}

\item{N_dist}{Claim sizes distribution c('pois', 'geom')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', 'invgauss')}

\item{beta}{Parameter in the weight function}

\item{B}{Number of loops in the parametric bootstrap}

\item{method}{Method of moment estimation c('MME', 'partial-MME')}

\item{scaled}{boolean to indicate wether the data is scaled}
}
\value{
the test statistic and the critrical value at a 95% confidence level
}
\description{
lt_L2_gof_pois
Compute the test statistic and critical value via a parametric bootstrap routine
for the LT based GOF test with a distance defined as an integrated square differrence
}
