% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_gof_CD.r
\name{lt_ODE_gof}
\alias{lt_ODE_gof}
\title{lt_ODE_gof
Compute the test statistic and the critical value of the test via a parametric bootstrap routine
when the distance is based on an ODE}
\usage{
lt_ODE_gof(X_data, N_dist, U_dist, beta, B, method = "partial-MME",
  scaled)
}
\arguments{
\item{X_data}{Sample of aggregated losses}

\item{N_dist}{Claim frequency distribution c('pois', 'geom')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', 'invgauss')}

\item{beta}{Parameter of the exponential weight function}

\item{B}{Number of loops in the parametric bootstrap routine}

\item{method}{Method of moment estimation method c('MME', 'partial-MME')}

\item{scaled}{Indicates wether or not the data is being scaled}
}
\value{
The value of the test statistic and the critical value at a 95% confidence level
}
\description{
lt_ODE_gof
Compute the test statistic and the critical value of the test via a parametric bootstrap routine
when the distance is based on an ODE
}
