% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_CD.r
\name{MME_CD}
\alias{MME_CD}
\title{MME_CD
Infer the parameter of a compound distribution}
\usage{
MME_CD(X_data, N_dist, U_dist, method = "partial-MME")
}
\arguments{
\item{X_data}{Sample of aggregated losses}

\item{N_dist}{Claim frequency distribution c('pois', 'geom')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', weibull', 'invgauss')}

\item{method}{MME of partial-MME (usimng the number of zeros in the data)}
}
\value{
the parameters of the compound distribution
}
\description{
MME_CD
Infer the parameter of a compound distribution
}
