% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_MC.r
\name{p_value}
\alias{p_value}
\title{p_value
Compute the p-value of a Compound distribution goodness-of-fit test using
the Warp speed method}
\usage{
p_value(n, N_dist_alt, U_dist_alt, par_alt_dist, N_dist_H0, U_dist_H0, MC,
  method_infer, method_gof, par_method)
}
\arguments{
\item{n}{Size of the sample}

\item{N_dist_alt}{Claim frequency distribution of the input sample}

\item{U_dist_alt}{Claim sizes distribution of the input sample}

\item{N_dist_H0}{Claim frequency distribution under H0}

\item{U_dist_H0}{Claim sizes distribution under H0}

\item{MC}{Number of Monte Carlo iteration}

\item{method_infer}{Type of MM estimation c('partial-MME', 'MME', )}

\item{method_gof}{Type of GOF test c('KS', 'CvM' 'LT_ODE', 'LT_L2')}

\item{par_method}{Parameter of the method discretization width for the distribution
function gof test and parameter of the weight function for the Laplace transform based
test c(h, beta)}
}
\value{
the p value of the test
}
\description{
p_value
Compute the p-value of a Compound distribution goodness-of-fit test using
the Warp speed method
}
