% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_CD.r
\name{Poisson_gamma_to_Poisson_exp}
\alias{Poisson_gamma_to_Poisson_exp}
\title{Poisson_gamma_to_Poisson_exp
Parameters of a Compound Poisson exponential model obtained by matching th emoment of the compound-Gamma model}
\usage{
Poisson_gamma_to_Poisson_exp(lam, r, theta)
}
\arguments{
\item{lam}{Intensity of the Poisson distribution}

\item{r}{shape of the gamma distribution}

\item{theta}{scale parameter of the gamma distribution}
}
\value{
The parameters of the Poisson exponential model
}
\description{
Poisson_gamma_to_Poisson_exp
Parameters of a Compound Poisson exponential model obtained by matching th emoment of the compound-Gamma model
}
