% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_CD.r
\name{simulate_X}
\alias{simulate_X}
\title{simulate_X
Generate a sample of aggregated claim sizes}
\usage{
simulate_X(n, N_dist, U_dist, X_par)
}
\arguments{
\item{n}{Size of the sample}

\item{N_dist}{Claim frequency distribution
c('pois', 'geom', 'dunif', 'nbinom', 'zmpois', 'mpois')}

\item{U_dist}{Claim sizes distribution c('exp', 'gamma', 'lnorm', 'weibull', 'invgauss')}

\item{X_par}{Parameters of the aggregated claim sizes distribution}
}
\value{
Sample of size n of the aggregated claim sizes
}
\description{
simulate_X
Generate a sample of aggregated claim sizes
}
