% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_CD.r
\name{sum_U}
\alias{sum_U}
\title{sum_U
Sum N iid observations drawn from the distribution U_dist}
\usage{
sum_U(N, U_dist, U_par)
}
\arguments{
\item{N}{Number of claim sizes}

\item{U_dist}{Distribution of the claim sizes c('exp', 'gamma', 'weibull', invgauss)}

\item{U_par}{Parameter of the distribution of the claim sizes}
}
\value{
The sum of N claim sizes
}
\description{
sum_U
Sum N iid observations drawn from the distribution U_dist
}
